Moved compute and storage classes to a separate file
