+++ pywren_ibm_cloud/runtime/extract_preinstalls_fn.py
